fix: Sling not respecting private_key property
When attempting to connect to snowflake using the snowflake_jwt
authenticator snowflake returns an error that the `private_key_path`
property is not set.

This changes it so that it will only error if both `private_key` and
`private_key_path` are not set.

This way existing users will not be effected by this change, and new
users are able to use the `private_key` property instead